% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hcp_file.R
\name{get_citations}
\alias{get_citations}
\alias{get_metadata}
\title{Fetch template citations}
\usage{
get_citations(template, as_bibtex = TRUE)

get_metadata(template)
}
\arguments{
\item{template}{A template identifier}

\item{as_bibtex}{The elements are returned as \code{\link{bibentry}}}
}
\value{
A list of references
}
\description{
Fetch template citations
}
