% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skeleton.R
\name{get_skeleton}
\alias{get_skeleton}
\alias{install_skeleton}
\alias{get_template_path}
\title{Get Skeleton File}
\usage{
get_skeleton()

install_skeleton(outdir = tf_home())

get_template_path()
}
\arguments{
\item{outdir}{output directory for skeleton}
}
\value{
A character file name
}
\description{
Get Skeleton File
}
\examples{
res = get_skeleton()
}
