% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hcp_file.R
\name{get_template_files}
\alias{get_template_files}
\alias{get_template}
\alias{list_template_files}
\alias{list_template_images}
\alias{templates}
\title{Get Template}
\usage{
get_template_files(
  files,
  outdir = NULL,
  verbose = TRUE,
  overwrite = FALSE,
  ...
)

get_template(template, outdir = NULL, verbose = TRUE, overwrite = FALSE, ...)

list_template_files(template)

list_template_images(template)

templates()
}
\arguments{
\item{files}{Path to file on S3 Bucket}

\item{outdir}{output directory for files to download}

\item{verbose}{should progress be added to downloading?}

\item{overwrite}{overwrite files if they already exist}

\item{...}{additional arguments to pass to other methods}

\item{template}{A template identifier}
}
\value{
Output filename character vector that was downloaded
}
\description{
Get Template
}
\examples{
temps = templates()
get_metadata(temps[1])
get_citations(temps[1])
list_template_files(temps[1])
all_files = list_template_files(temps)
out = list_template_images(temps[1])
data = get_template_files(out, outdir = tempfile())
}
