% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{print.TemplateFlowClient}
\alias{print.TemplateFlowClient}
\title{Print a TemplateFlowClient}
\usage{
\method{print}{TemplateFlowClient}(x, ...)
}
\arguments{
\item{x}{A `TemplateFlowClient` object.}

\item{...}{Additional arguments (ignored).}
}
\value{
`x`, invisibly.
}
\description{
Print a TemplateFlowClient
}
