% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tf_cache_stats}
\alias{tf_cache_stats}
\title{Cache statistics}
\usage{
tf_cache_stats(cache = NULL)
}
\arguments{
\item{cache}{A `TemplateFlowCache` object or a character path to the cache
root (optional; uses default client cache).}
}
\value{
A list with `cache_files`, `cache_bytes`, and `cache_size_human`.
}
\description{
Report the number and total size of cached files.
}
\examples{
\dontrun{
tf_cache_stats()
}
}
