% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{tf_client}
\alias{tf_client}
\title{Get or create the default global TemplateFlow client}
\usage{
tf_client()
}
\value{
A `TemplateFlowClient`.
}
\description{
Get or create the default global TemplateFlow client
}
