% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalad.R
\name{tf_datalad_get}
\alias{tf_datalad_get}
\title{Get files from a DataLad dataset}
\usage{
tf_datalad_get(files, dataset)
}
\arguments{
\item{files}{Character vector of file paths to retrieve.}

\item{dataset}{Path to the DataLad dataset root.}
}
\value{
Invisible files vector.
}
\description{
Retrieve (download) specific files from a DataLad dataset.
}
\examples{
\dontrun{
tf_datalad_get(c("tpl-MNI152Lin/tpl-MNI152Lin_res-01_T1w.nii.gz"),
  dataset = "~/.cache/templateflow")
}
}
