% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalad.R
\name{tf_datalad_install}
\alias{tf_datalad_install}
\title{Install a DataLad dataset}
\usage{
tf_datalad_install(path, source, recursive = TRUE)
}
\arguments{
\item{path}{Local destination path.}

\item{source}{Dataset source URL or path.}

\item{recursive}{If TRUE, install subdatasets recursively.}
}
\value{
Invisible path.
}
\description{
Clone/install a DataLad dataset to a local path.
}
\examples{
\dontrun{
tf_datalad_install("~/.cache/templateflow",
  source = "https://github.com/templateflow/templateflow.git")
}
}
