% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalad.R
\name{tf_datalad_update}
\alias{tf_datalad_update}
\title{Update a DataLad dataset}
\usage{
tf_datalad_update(dataset, recursive = TRUE, merge = TRUE)
}
\arguments{
\item{dataset}{Path to the DataLad dataset root.}

\item{recursive}{If TRUE, update subdatasets recursively.}

\item{merge}{If TRUE, merge fetched changes.}
}
\value{
Invisible TRUE.
}
\description{
Fetch updates from the remote and optionally merge.
}
\examples{
\dontrun{
tf_datalad_update("~/.cache/templateflow", recursive = TRUE, merge = TRUE)
}
}
