% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{tf_file_extension}
\alias{tf_file_extension}
\title{Extract compound file extension}
\usage{
tf_file_extension(path)
}
\arguments{
\item{path}{File path.}
}
\value{
The file extension including leading dot.
}
\description{
Handles compound extensions like `.nii.gz`, `.surf.gii`, etc.
}
