% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{tf_get_citations}
\alias{tf_get_citations}
\title{Retrieve template citations}
\usage{
tf_get_citations(client = NULL, template, bibtex = FALSE)
}
\arguments{
\item{client}{A `TemplateFlowClient` (optional; default global client).}

\item{template}{Template identifier (e.g., `"MNI152Lin"`).}

\item{bibtex}{If TRUE, fetch citations in BibTeX format.}
}
\value{
Character vector of citation URLs or BibTeX entries.
}
\description{
Retrieve template citations
}
\examples{
\dontrun{
tf_get_citations(template = "MNI152NLin2009cAsym")
tf_get_citations(template = "MNI152NLin2009cAsym", bibtex = TRUE)
}
}
