% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{tf_get_metadata}
\alias{tf_get_metadata}
\title{Read template metadata}
\usage{
tf_get_metadata(client = NULL, template)
}
\arguments{
\item{client}{A `TemplateFlowClient` (optional; default global client).}

\item{template}{Template identifier (e.g., `"MNI152Lin"`).}
}
\value{
List parsed from `template_description.json`.
}
\description{
Read template metadata
}
\examples{
\dontrun{
meta <- tf_get_metadata(template = "MNI152Lin")
meta$Name
}
}
