% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{tf_set_client}
\alias{tf_set_client}
\title{Set global client}
\usage{
tf_set_client(client)
}
\arguments{
\item{client}{A `TemplateFlowClient`.}
}
\value{
The client (invisibly).
}
\description{
Set global client
}
\examples{
\dontrun{
client <- TemplateFlowClient()
tf_set_client(client)
}
}
